add new plugin file for building spotlight index page
