+++ website/docusaurus.config.js
@@ -228,6 +228,7 @@ var siteSettings = {
     path.resolve("plugins/customWebpackConfig"),
     [path.resolve("plugins/buildGlobalData"), { versionedPages }],
     path.resolve("plugins/buildAuthorPages"),
+    path.resolve("plugins/buildSpotlightIndexPage"),
   ],
   scripts: [
     {
